requirements.txt
django-model-utils>=2.0
pillow
# Additional requirements go here
django-imagekit
